Ifdef things so they can build against Dev14.
