+++ src/EditorFeatures/Core/Implementation/IntelliSense/Completion/Temporary/IIntellisenseFilter.cs
@@ -1,4 +1,4 @@
-﻿#if false
+﻿#if !NEWCOMPLETION
 // Copyright (c) Microsoft Corporation
 // All rights reserved
 // REMOVE ONCE WE ACTUALLY REFERENCE THE REAL EDITOR DLLS.
